Add src/survey_scoring.js
Add a file containing the question scoring system. The questions'
answers' arrays of numbers represent the scores given to different study
programs. First number is the ID of the program and the second one is
the score given (mainly +1 or 0 i.e., none). The 0's could be omitted
but I left them there for demonstration purposes.