Config for CSC Rahti OpenShift deployment
Configuration for deploying the application using Kubernetes to CSC
Rahti OpenShift. In addition, the Apache2 web server is switched for
Nginx so that the OpenShift security policies needn't be dealt with.

By default, OpenShift uses random UIDs for the pods' services and all
files are owned by root. As such, Apache2 does not have access to modify
its lock file in /var/lock/, which causes it to not run happily.