Refactor and extend database schema
Add specific ID column to the study_programs table. Traditionally, the
DBID column would be used for referencing between tables and this new
column is a bit redundant, but seeing as the DBID is not as easily
modifiable, and the frontend configuration depends on reliable ID
numbering, it is likely a good idea to separate this from DBID.

Rename the respondents table to survey_answers, and include more
optional fields (as well as the mandatory page arrival timestamp).